More 1.20.2 work

+++ src/main/java/com/simibubi/create/content/fluids/particle/FluidStackParticle.java
@@ -69,19 +69,19 @@ public class FluidStackParticle extends TextureSheetParticle {
 	}
 
 	protected float getU0() {
-		return this.sprite.getU((double) ((this.uo + 1.0F) / 4.0F * 16.0F));
+		return this.sprite.getU((this.uo + 1.0F) / 4.0F * 16.0F);
 	}
 
 	protected float getU1() {
-		return this.sprite.getU((double) (this.uo / 4.0F * 16.0F));
+		return this.sprite.getU(this.uo / 4.0F * 16.0F);
 	}
 
 	protected float getV0() {
-		return this.sprite.getV((double) (this.vo / 4.0F * 16.0F));
+		return this.sprite.getV(this.vo / 4.0F * 16.0F);
 	}
 
 	protected float getV1() {
-		return this.sprite.getV((double) ((this.vo + 1.0F) / 4.0F * 16.0F));
+		return this.sprite.getV((this.vo + 1.0F) / 4.0F * 16.0F);
 	}
 
 	@Override
